+++ hdl/common/include/util/memory_mapping.hpp
@@ -50,11 +50,11 @@ namespace MemoryMapping {
 
 }
 
-#include "memory_mapping/register.hpp"
-#include "memory_mapping/config.hpp"
-#include "memory_mapping/field.hpp"
-#include "memory_mapping/bitfield.hpp"
-#include "memory_mapping/virtual_field.hpp"
-#include "memory_mapping/virtual_config.hpp"
+#include "register.hpp"
+#include "config.hpp"
+#include "field.hpp"
+#include "bitfield.hpp"
+#include "virtual_field.hpp"
+#include "virtual_config.hpp"
 
 #endif//MEMORY_MAPPING_HPP
