+++ hdl/common/include/util/virtual_field.hpp
